Crate cmdtaglib

A command line application that allows you to interact with taglib. It allows for easy reading and writing of tags for a script or user.

Modules

cmd
Provides the CMD enum, representing the different commands that can be invoked.

help
Provides the show_help(&String) function, which shows help for a user who incorrectly used the program.

parse
Provides the parse_cmd(&String, &str) function.